Extract domain-specific configurations from src/config.rs into their respective submodules.
Refactored `src/config.rs` from a "Blob" module into a clean Facade. Extracted `WalConfig`, `HistoricalConfig`, and `VectorIndexConfig` (and their builders) into `src/storage/wal/config.rs`, `src/storage/historical/config.rs`, and `src/index/vector/config.rs` respectively. Added `Default` implementations for the builders to resolve clippy warnings. Updated module exports to present a unified configuration API while enforcing strict domain boundaries.
