+++ src/config.rs
@@ -44,663 +44,10 @@ use std::fs;
 use std::path::Path;
 
 use crate::storage::index_persistence::PersistenceConfig;
-use crate::storage::version::AnchorConfig;
 
-/// Configuration for WAL (Write-Ahead Log) system.
-///
-/// Controls buffer sizes, stripe configuration, flush behavior, and durability settings.
-/// This consolidates all WAL-related configuration in one place.
-#[derive(Debug, Clone, PartialEq)]
-#[cfg_attr(feature = "config-toml", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "config-toml", serde(default))]
-#[non_exhaustive]
-pub struct WalConfig {
-    /// Number of stripes for concurrent appends (must be power of 2).
-    /// Higher values improve concurrency but use more memory.
-    /// Default: 16
-    pub num_stripes: usize,
-
-    /// Ring buffer capacity per stripe.
-    /// Larger capacity reduces backpressure but uses more memory.
-    /// Default: 1024
-    pub stripe_capacity: usize,
-
-    /// Write buffer size for segment files (in bytes).
-    /// Affects I/O batching efficiency.
-    /// Default: 64KB (65536 bytes)
-    pub write_buffer_size: usize,
-
-    /// Maximum segment size before rotation (in bytes).
-    /// Default: 64MB (67108864 bytes)
-    pub segment_size: usize,
-
-    /// Flush interval in milliseconds for async/group-commit modes.
-    /// Lower values reduce latency but increase I/O overhead.
-    /// Default: 10ms
-    pub flush_interval_ms: u64,
-
-    /// Directory where WAL files are stored.
-    /// Default: "aletheiadb/wal"
-    pub wal_dir: std::path::PathBuf,
-
-    /// Number of WAL segments to keep for recovery.
-    /// Default: 10
-    pub segments_to_retain: usize,
-
-    /// Durability mode controlling when data is synced to disk.
-    /// This determines the tradeoff between durability guarantees and performance.
-    /// Default: GroupCommit (10ms delay, 200 batch size)
-    pub durability_mode: crate::storage::wal::DurabilityMode,
-}
-
-impl Default for WalConfig {
-    fn default() -> Self {
-        Self {
-            num_stripes: 16,
-            stripe_capacity: 1024,
-            write_buffer_size: 64 * 1024,   // 64KB
-            segment_size: 64 * 1024 * 1024, // 64MB
-            flush_interval_ms: 10,
-            wal_dir: std::path::PathBuf::from("aletheiadb/wal"),
-            segments_to_retain: 10,
-            durability_mode: crate::storage::wal::DurabilityMode::group_commit_default(),
-        }
-    }
-}
-
-/// Builder for WAL configuration.
-///
-/// Provides a fluent API for constructing WAL configuration with validation.
-#[must_use = "builders do nothing unless you call build()"]
-#[derive(Debug)]
-pub struct WalConfigBuilder {
-    config: WalConfig,
-}
-
-impl WalConfigBuilder {
-    /// Create a new builder with default values.
-    pub fn new() -> Self {
-        Self {
-            config: WalConfig::default(),
-        }
-    }
-
-    /// Set all validated parameters at once (single validation point).
-    ///
-    /// This is a convenience method that sets all parameters requiring validation
-    /// in a single call, reducing the need for multiple `.unwrap()` calls.
-    ///
-    /// # Parameters
-    ///
-    /// - `num_stripes`: Number of stripes for concurrent appends (will be rounded to next power of 2)
-    /// - `stripe_capacity`: Ring buffer capacity per stripe
-    /// - `write_buffer_size`: Write buffer size in bytes
-    /// - `segment_size`: Maximum segment size before rotation in bytes (minimum 512)
-    /// - `segments_to_retain`: Number of WAL segments to keep for recovery
-    /// - `flush_interval_ms`: Flush interval in milliseconds for async/group-commit modes
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if any parameter is invalid:
-    /// - Any value is 0
-    /// - `segment_size` is less than 512 bytes
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use aletheiadb::WalConfigBuilder;
-    ///
-    /// let config = WalConfigBuilder::new()
-    ///     .with_validated(
-    ///         32,              // num_stripes
-    ///         2048,            // stripe_capacity
-    ///         128 * 1024,      // write_buffer_size
-    ///         64 * 1024 * 1024, // segment_size
-    ///         10,              // segments_to_retain
-    ///         10,              // flush_interval_ms
-    ///     ).unwrap()  // Single unwrap!
-    ///     .build();
-    /// ```
-    pub fn with_validated(
-        self,
-        num_stripes: usize,
-        stripe_capacity: usize,
-        write_buffer_size: usize,
-        segment_size: usize,
-        segments_to_retain: usize,
-        flush_interval_ms: u64,
-    ) -> Result<Self, ConfigError> {
-        self.num_stripes(num_stripes)?
-            .stripe_capacity(stripe_capacity)?
-            .write_buffer_size(write_buffer_size)?
-            .segment_size(segment_size)?
-            .segments_to_retain(segments_to_retain)?
-            .flush_interval_ms(flush_interval_ms)
-    }
-
-    /// Set the number of stripes (will be rounded to next power of 2).
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `num_stripes` is 0.
-    pub fn num_stripes(mut self, num_stripes: usize) -> Result<Self, ConfigError> {
-        if num_stripes == 0 {
-            return Err(ConfigError::InvalidValue(
-                "num_stripes must be greater than 0".into(),
-            ));
-        }
-        let rounded = num_stripes.next_power_of_two();
-        if rounded != num_stripes {
-            #[cfg(feature = "observability")]
-            tracing::warn!(
-                original = num_stripes,
-                rounded = rounded,
-                "num_stripes rounded to next power of 2"
-            );
-        }
-        self.config.num_stripes = rounded;
-        Ok(self)
-    }
-
-    /// Set the stripe capacity.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `capacity` is 0.
-    pub fn stripe_capacity(mut self, capacity: usize) -> Result<Self, ConfigError> {
-        if capacity == 0 {
-            return Err(ConfigError::InvalidValue(
-                "stripe_capacity must be greater than 0".into(),
-            ));
-        }
-        self.config.stripe_capacity = capacity;
-        Ok(self)
-    }
-
-    /// Set the write buffer size in bytes.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `size` is 0.
-    pub fn write_buffer_size(mut self, size: usize) -> Result<Self, ConfigError> {
-        if size == 0 {
-            return Err(ConfigError::InvalidValue(
-                "write_buffer_size must be greater than 0".into(),
-            ));
-        }
-        self.config.write_buffer_size = size;
-        Ok(self)
-    }
-
-    /// Set the segment size in bytes.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `size` is 0 or less than 512 bytes.
-    ///
-    /// **Note**: While 512 bytes is allowed (for testing), production use should
-    /// be at least 1MB for reasonable performance.
-    pub fn segment_size(mut self, size: usize) -> Result<Self, ConfigError> {
-        const MIN_SEGMENT_SIZE: usize = 512; // Allow small sizes for testing
-        if size == 0 {
-            return Err(ConfigError::InvalidValue(
-                "segment_size must be greater than 0".into(),
-            ));
-        }
-        if size < MIN_SEGMENT_SIZE {
-            return Err(ConfigError::InvalidValue(format!(
-                "segment_size must be at least {} bytes, got {}",
-                MIN_SEGMENT_SIZE, size
-            )));
-        }
-        self.config.segment_size = size;
-        Ok(self)
-    }
-
-    /// Set the flush interval in milliseconds.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `ms` is 0.
-    pub fn flush_interval_ms(mut self, ms: u64) -> Result<Self, ConfigError> {
-        if ms == 0 {
-            return Err(ConfigError::InvalidValue(
-                "flush_interval_ms must be greater than 0".into(),
-            ));
-        }
-        self.config.flush_interval_ms = ms;
-        Ok(self)
-    }
-
-    /// Set the WAL directory path.
-    pub fn wal_dir(mut self, path: std::path::PathBuf) -> Self {
-        self.config.wal_dir = path;
-        self
-    }
-
-    /// Set the number of WAL segments to retain.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `segments_to_retain` is 0.
-    pub fn segments_to_retain(mut self, segments: usize) -> Result<Self, ConfigError> {
-        if segments == 0 {
-            return Err(ConfigError::InvalidValue(
-                "segments_to_retain must be greater than 0".into(),
-            ));
-        }
-        self.config.segments_to_retain = segments;
-        Ok(self)
-    }
-
-    /// Set the durability mode.
-    pub fn durability_mode(mut self, mode: crate::storage::wal::DurabilityMode) -> Self {
-        self.config.durability_mode = mode;
-        self
-    }
-
-    /// Build the configuration.
-    pub fn build(self) -> WalConfig {
-        self.config
-    }
-}
-
-impl Default for WalConfigBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Configuration for historical storage.
-///
-/// Controls versioning, reconstruction limits, and caching behavior.
-#[derive(Debug, Clone, PartialEq)]
-#[cfg_attr(feature = "config-toml", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "config-toml", serde(default))]
-#[non_exhaustive]
-pub struct HistoricalConfig {
-    /// Maximum versions to retain per entity before pruning.
-    /// Higher values preserve more history but use more memory.
-    /// Default: 1000
-    pub max_versions_per_entity: usize,
-
-    /// Maximum depth for delta chain reconstruction.
-    /// Protects against stack overflow and infinite loops.
-    /// Default: 100
-    pub max_reconstruction_depth: usize,
-
-    /// Size of the reconstruction cache (number of entries).
-    /// Larger cache improves temporal query performance.
-    /// Default: 10000
-    pub reconstruction_cache_size: usize,
-
-    /// Create an anchor every N versions (default: 10).
-    /// Smaller intervals mean faster reconstruction but more storage.
-    pub anchor_interval: u32,
-
-    /// Maximum delta chain length before forcing an anchor (default: 20).
-    /// Ensures reconstruction cost is bounded.
-    pub max_delta_chain: u32,
-
-    /// Enable cold storage (Redb-based tiered storage) for unlimited historical depth.
-    /// When enabled, old versions are migrated to disk automatically.
-    /// Default: false
-    pub enable_cold_storage: bool,
-
-    /// Path to the cold storage Redb file.
-    /// Required if `enable_cold_storage` is true.
-    /// Default: None
-    pub cold_storage_path: Option<std::path::PathBuf>,
-
-    /// Age threshold for migrating versions to cold storage.
-    /// Versions older than this duration are eligible for migration.
-    /// Default: 1 hour
-    pub migration_age_threshold: std::time::Duration,
-
-    /// Maximum number of hot versions to keep per entity before triggering migration.
-    /// Default: 1000 (same as max_versions_per_entity)
-    pub max_hot_versions: usize,
-}
-
-impl Default for HistoricalConfig {
-    fn default() -> Self {
-        let anchor_defaults = AnchorConfig::default();
-        Self {
-            max_versions_per_entity: 1000,
-            max_reconstruction_depth: 100,
-            reconstruction_cache_size: 10000,
-            anchor_interval: anchor_defaults.anchor_interval,
-            max_delta_chain: anchor_defaults.max_delta_chain,
-            enable_cold_storage: false,
-            cold_storage_path: None,
-            migration_age_threshold: std::time::Duration::from_secs(3600), // 1 hour
-            max_hot_versions: 1000,
-        }
-    }
-}
-
-/// Builder for historical storage configuration.
-///
-/// Provides a fluent API for constructing historical storage configuration with validation.
-#[must_use = "builders do nothing unless you call build()"]
-#[derive(Debug)]
-pub struct HistoricalConfigBuilder {
-    config: HistoricalConfig,
-}
-
-impl HistoricalConfigBuilder {
-    /// Create a new builder with default values.
-    pub fn new() -> Self {
-        Self {
-            config: HistoricalConfig::default(),
-        }
-    }
-
-    /// Set the maximum versions per entity.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `max` is 0.
-    pub fn max_versions_per_entity(mut self, max: usize) -> Result<Self, ConfigError> {
-        if max == 0 {
-            return Err(ConfigError::InvalidValue(
-                "max_versions_per_entity must be greater than 0".into(),
-            ));
-        }
-        self.config.max_versions_per_entity = max;
-        Ok(self)
-    }
-
-    /// Set the maximum reconstruction depth.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `depth` is 0 or greater than 1000.
-    pub fn max_reconstruction_depth(mut self, depth: usize) -> Result<Self, ConfigError> {
-        if depth == 0 {
-            return Err(ConfigError::InvalidValue(
-                "max_reconstruction_depth must be greater than 0".into(),
-            ));
-        }
-        if depth > 1000 {
-            return Err(ConfigError::InvalidValue(
-                "max_reconstruction_depth cannot exceed 1000 (risk of stack overflow)".into(),
-            ));
-        }
-        self.config.max_reconstruction_depth = depth;
-        Ok(self)
-    }
-
-    /// Set the reconstruction cache size.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `size` is 0.
-    pub fn reconstruction_cache_size(mut self, size: usize) -> Result<Self, ConfigError> {
-        if size == 0 {
-            return Err(ConfigError::InvalidValue(
-                "reconstruction_cache_size must be greater than 0".into(),
-            ));
-        }
-        self.config.reconstruction_cache_size = size;
-        Ok(self)
-    }
-
-    /// Set the anchor interval.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `interval` is 0.
-    pub fn anchor_interval(mut self, interval: u32) -> Result<Self, ConfigError> {
-        if interval == 0 {
-            return Err(ConfigError::InvalidValue(
-                "anchor_interval must be greater than 0".into(),
-            ));
-        }
-        self.config.anchor_interval = interval;
-        Ok(self)
-    }
-
-    /// Set the maximum delta chain length.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `max` is 0.
-    pub fn max_delta_chain(mut self, max: u32) -> Result<Self, ConfigError> {
-        if max == 0 {
-            return Err(ConfigError::InvalidValue(
-                "max_delta_chain must be greater than 0".into(),
-            ));
-        }
-        self.config.max_delta_chain = max;
-        Ok(self)
-    }
-
-    /// Enable cold storage (Redb-based tiered storage).
-    ///
-    /// When enabled, old versions are automatically migrated to disk, allowing
-    /// unlimited historical depth without consuming RAM.
-    ///
-    /// # Note
-    ///
-    /// You must also call [`cold_storage_path`](Self::cold_storage_path) to specify
-    /// where the Redb file should be stored, or the database will fail to initialize.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use aletheiadb::config::HistoricalConfigBuilder;
-    ///
-    /// let config = HistoricalConfigBuilder::new()
-    ///     .enable_cold_storage(true)
-    ///     .cold_storage_path("data/cold.redb")
-    ///     .build();
-    /// ```
-    pub fn enable_cold_storage(mut self, enabled: bool) -> Self {
-        self.config.enable_cold_storage = enabled;
-        self
-    }
-
-    /// Set the path to the cold storage Redb file.
-    ///
-    /// This is required if [`enable_cold_storage`](Self::enable_cold_storage) is true.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use aletheiadb::config::HistoricalConfigBuilder;
-    ///
-    /// let config = HistoricalConfigBuilder::new()
-    ///     .enable_cold_storage(true)
-    ///     .cold_storage_path("data/cold.redb")
-    ///     .build();
-    /// ```
-    pub fn cold_storage_path<P: Into<std::path::PathBuf>>(mut self, path: P) -> Self {
-        self.config.cold_storage_path = Some(path.into());
-        self
-    }
-
-    /// Set the age threshold for migrating versions to cold storage.
-    ///
-    /// Versions older than this duration become eligible for migration to disk.
-    ///
-    /// # Default
-    ///
-    /// 1 hour (3600 seconds)
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use aletheiadb::config::HistoricalConfigBuilder;
-    /// use std::time::Duration;
-    ///
-    /// let config = HistoricalConfigBuilder::new()
-    ///     .enable_cold_storage(true)
-    ///     .cold_storage_path("data/cold.redb")
-    ///     .migration_age_threshold(Duration::from_secs(7200)) // 2 hours
-    ///     .build();
-    /// ```
-    pub fn migration_age_threshold(mut self, threshold: std::time::Duration) -> Self {
-        self.config.migration_age_threshold = threshold;
-        self
-    }
-
-    /// Set the maximum number of hot versions to keep per entity.
-    ///
-    /// When the number of versions exceeds this threshold, older versions
-    /// are migrated to cold storage.
-    ///
-    /// # Default
-    ///
-    /// 1000 (same as `max_versions_per_entity`)
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use aletheiadb::config::HistoricalConfigBuilder;
-    ///
-    /// let config = HistoricalConfigBuilder::new()
-    ///     .enable_cold_storage(true)
-    ///     .cold_storage_path("data/cold.redb")
-    ///     .max_hot_versions(500) // Keep only 500 versions in RAM
-    ///     .build();
-    /// ```
-    pub fn max_hot_versions(mut self, max: usize) -> Self {
-        self.config.max_hot_versions = max;
-        self
-    }
-
-    /// Build the configuration.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `enable_cold_storage` is true but `cold_storage_path` is not set.
-    /// Use [`build_checked`](Self::build_checked) for a non-panicking version.
-    pub fn build(self) -> HistoricalConfig {
-        if self.config.enable_cold_storage && self.config.cold_storage_path.is_none() {
-            panic!("cold_storage_path must be set when enable_cold_storage is true");
-        }
-        self.config
-    }
-
-    /// Build the configuration with validation.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `enable_cold_storage` is true
-    /// but `cold_storage_path` is not set.
-    pub fn build_checked(self) -> Result<HistoricalConfig, ConfigError> {
-        if self.config.enable_cold_storage && self.config.cold_storage_path.is_none() {
-            return Err(ConfigError::InvalidValue(
-                "cold_storage_path must be set when enable_cold_storage is true".into(),
-            ));
-        }
-        Ok(self.config)
-    }
-}
-
-impl Default for HistoricalConfigBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Configuration for vector index system.
-///
-/// Controls limits for k-NN queries and HNSW index structure.
-#[derive(Debug, Clone, PartialEq)]
-#[cfg_attr(feature = "config-toml", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "config-toml", serde(default))]
-#[non_exhaustive]
-pub struct VectorIndexConfig {
-    /// Maximum value of k for k-NN queries.
-    /// Prevents excessive memory usage from large result sets.
-    /// Default: 10000
-    pub max_k: usize,
-
-    /// Maximum layer depth in HNSW index.
-    /// Controls the maximum graph structure depth.
-    /// Default: 16
-    pub max_layer: usize,
-}
-
-impl Default for VectorIndexConfig {
-    fn default() -> Self {
-        Self {
-            max_k: 10000,
-            max_layer: 16,
-        }
-    }
-}
-
-/// Builder for vector index configuration.
-///
-/// Provides a fluent API for constructing vector index configuration with validation.
-#[must_use = "builders do nothing unless you call build()"]
-#[derive(Debug)]
-pub struct VectorIndexConfigBuilder {
-    config: VectorIndexConfig,
-}
-
-impl VectorIndexConfigBuilder {
-    /// Create a new builder with default values.
-    pub fn new() -> Self {
-        Self {
-            config: VectorIndexConfig::default(),
-        }
-    }
-
-    /// Set the maximum k for k-NN queries.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `k` is 0 or greater than 100,000.
-    pub fn max_k(mut self, k: usize) -> Result<Self, ConfigError> {
-        if k == 0 {
-            return Err(ConfigError::InvalidValue(
-                "max_k must be greater than 0".into(),
-            ));
-        }
-        if k > 100_000 {
-            return Err(ConfigError::InvalidValue(
-                "max_k cannot exceed 100,000 (DoS protection)".into(),
-            ));
-        }
-        self.config.max_k = k;
-        Ok(self)
-    }
-
-    /// Set the maximum layer depth.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConfigError::InvalidValue` if `layer` is 0 or greater than 32.
-    pub fn max_layer(mut self, layer: usize) -> Result<Self, ConfigError> {
-        if layer == 0 {
-            return Err(ConfigError::InvalidValue(
-                "max_layer must be greater than 0".into(),
-            ));
-        }
-        if layer > 32 {
-            return Err(ConfigError::InvalidValue(
-                "max_layer cannot exceed 32 (HNSW limitation)".into(),
-            ));
-        }
-        self.config.max_layer = layer;
-        Ok(self)
-    }
-
-    /// Build the configuration.
-    pub fn build(self) -> VectorIndexConfig {
-        self.config
-    }
-}
-
-impl Default for VectorIndexConfigBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
+pub use crate::index::vector::config::{VectorIndexConfig, VectorIndexConfigBuilder};
+pub use crate::storage::historical::config::{HistoricalConfig, HistoricalConfigBuilder};
+pub use crate::storage::wal::config::{WalConfig, WalConfigBuilder};
 
 /// Unified configuration for AletheiaDB.
 ///
